Added a feature to make soulsand renewable.
*read the following with a mythical explainer voice*

To create soul sand you need to burn a soul of a living entity into a block of sand.
So it is necessary to have a sand block as base on top of that you'll need a fire block
and last a living entity must be sacrificed on that combination only then it's soul will
be burned into the sand and become a soul sand.

The rule for this is "mobInFireConvertsSandToSoulsand".

I hope you like the description :-)